Fix bug #95772

+++ nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kmf/004/d/kmf004.d.vm.ts
@@ -298,6 +298,11 @@ module nts.uk.at.view.kmf004.d.viewmodel {
                 }
             }
             
+            if(self.provisionCheck()) {
+                nts.uk.ui.dialog.alertError({ messageId: "Msg_1219" });
+                return;
+            }
+            
             nts.uk.ui.dialog.confirm({ messageId: "Msg_18" }).ifYes(() => { 
                 service.remove(self.selectedOption()).done(function(){
                     self.getData().done(function(){
